# docs/swagger.yaml
definitions:
...
  controller.SaveModelConfigsRequest:
    properties:
      cache_creation_price:
        type: number
      cached_price:
        type: number
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      exclude_from_tests:
        type: boolean
      image_batch_size:
        type: integer
      image_prices:
        additionalProperties:
          type: number
        type: object
      input_price:
        type: number
      model:
        type: string
      output_price:
        type: number
      owner:
        $ref: '#/definitions/model.ModelOwner'
      rpm:
        type: integer
      tpm:
        type: integer
      type:
        $ref: '#/definitions/relaymode.Mode'
      updated_at:
        type: string
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.ModelConfig:
    properties:
      cache_creation_price:
        type: number
      cached_price:
        type: number
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      exclude_from_tests:
        type: boolean
      image_batch_size:
        type: integer
      image_prices:
        additionalProperties:
          type: number
        type: object
      input_price:
        type: number
      model:
        type: string
      output_price:
        type: number
      owner:
        $ref: '#/definitions/model.ModelOwner'
      rpm:
        type: integer
      tpm:
        type: integer
      type:
        $ref: '#/definitions/relaymode.Mode'
      updated_at:
        type: string
    type: object
...
  model.ModelOwner:
    enum:
    - openai
    - alibaba
    - tencent
    - xunfei
    - deepseek
    - moonshot
    - minimax
    - baidu
    - google
    - baai
    - funaudiollm
    - doubao
    - fishaudio
    - chatglm
    - stabilityai
    - netease
    - ai360
    - anthropic
    - meta
    - baichuan
    - mistral
    - openchat
    - microsoft
    - defog
    - nexusflow
    - cohere
    - huggingface
    - lingyiwanwu
    - stepfun
    - xai
    - doc2x
    type: string
    x-enum-varnames:
    - ModelOwnerOpenAI
    - ModelOwnerAlibaba
    - ModelOwnerTencent
    - ModelOwnerXunfei
    - ModelOwnerDeepSeek
    - ModelOwnerMoonshot
    - ModelOwnerMiniMax
    - ModelOwnerBaidu
    - ModelOwnerGoogle
    - ModelOwnerBAAI
    - ModelOwnerFunAudioLLM
    - ModelOwnerDoubao
    - ModelOwnerFishAudio
    - ModelOwnerChatGLM
    - ModelOwnerStabilityAI
    - ModelOwnerNetease
    - ModelOwnerAI360
    - ModelOwnerAnthropic
    - ModelOwnerMeta
    - ModelOwnerBaichuan
    - ModelOwnerMistral
    - ModelOwnerOpenChat
    - ModelOwnerMicrosoft
    - ModelOwnerDefog
    - ModelOwnerNexusFlow
    - ModelOwnerCohere
    - ModelOwnerHuggingFace
    - ModelOwnerLingyiWanwu
    - ModelOwnerStepFun
    - ModelOwnerXAI
    - ModelOwnerDoc2x
...
  relaymode.Mode:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    type: integer
    x-enum-varnames:
    - Unknown
    - ChatCompletions
    - Completions
    - Embeddings
    - Moderations
    - ImagesGenerations
    - Edits
    - AudioSpeech
    - AudioTranscription
    - AudioTranslation
    - Rerank
    - ParsePdf
info:
  contact: {}
paths:
...
  /api/model_config:
    post:
      description: Saves a model config
      parameters:
      - description: Model config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/controller.SaveModelConfigsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Save model config
      tags:
      - modelconfig
  /api/model_config/{model}:
    delete:
      description: Deletes a model config
      parameters:
      - description: Model name
        in: path
        name: model
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete model config
      tags:
      - modelconfig
    get:
      description: Returns a model config
      parameters:
      - description: Model name
        in: path
        name: model
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ModelConfig'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get model config
      tags:
      - modelconfig
  /api/model_configs:
    post:
      description: Saves a list of model configs
      parameters:
      - description: Model configs
        in: body
        name: configs
        required: true
        schema:
          items:
            $ref: '#/definitions/controller.SaveModelConfigsRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Save model configs
      tags:
      - modelconfig
  /api/model_configs/batch_delete:
    post:
      description: Deletes a list of model configs
      parameters:
      - description: Model names
        in: body
        name: models
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete model configs
      tags:
      - modelconfig
...
  /api/modelconfigs:
    get:
      description: Returns a list of model configs with pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        configs:
                          items:
                            $ref: '#/definitions/model.ModelConfig'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get model configs
      tags:
      - modelconfig
  /api/modelconfigs/all:
    get:
      description: Returns a list of all model configs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all model configs
      tags:
      - modelconfig
  /api/modelconfigs/contains:
    post:
      description: Returns a list of model configs by models contains
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get model configs by models contains
      tags:
      - modelconfig
  /api/modelconfigs/search:
    get:
      description: Returns a list of model configs by keyword
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        configs:
                          items:
                            $ref: '#/definitions/model.ModelConfig'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Search model configs
      tags:
      - modelconfig
...
  /api/monitor:
    delete:
      description: Clears all model errors
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Clear all model errors
      tags:
      - monitor
    get:
      description: Returns a list of all channel model error rates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    additionalProperties:
                      type: number
                    type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all channel model error rates
      tags:
      - monitor
  /api/monitor/{id}:
    delete:
      description: Clears all model errors for a specific channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Clear channel all model errors
      tags:
      - monitor
    get:
      description: Returns a list of channel model error rates
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    additionalProperties:
                      type: number
                    type: object
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel model error rates
      tags:
      - monitor
  /api/monitor/{id}/{model}:
    delete:
      description: Clears model errors for a specific channel and model
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Model name
        in: path
        name: model
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Clear channel model errors
      tags:
      - monitor
  /api/monitor/banned_channels:
    get:
      description: Returns a list of all banned model channels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    items:
                      type: integer
                    type: array
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all banned model channels
      tags:
      - monitor
  /api/monitor/models:
    get:
      description: Returns a list of models error rate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: number
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get models error rate
      tags:
      - monitor
...
swagger: "2.0"